// apps/trade-pro/src/pages/customerPortal/customerReports/SaleOrderReport/types.ts
export type TSaleOrderReport = {
  Id: number;
  PoDId: number;
  DocumentTypeId: number;
  DocumentTypeDescription: string;
  DocNo: number;
  DocDate: string;
  CustomerName: string;
  OrderDueDate: string;
  OrderDueDays: number;
  OrderStatus: string;
  SupplierRefNo: string;
  TermsDescription: string;
  OrderExpiryDate: string;
  DeliveryTerm: string;
  DeliveryDays: number;
  DeliveryRemarks: string | null;
  RemarksHeader: string;
  UserNameEusr: string;
  EntryDate: string;
  UserNameMusr: string | null;
  ModifyDate: string | null;
  UserNameAusr: string | null;
  PostDate: string;
  ItemCode: string;
  ItemName: string;
  Crop: string;
  OrderItemQty: number;
  DispatchQty: number;
  BalQty: number;
  NetWeight: number;
  DispatchWeight: number;
  BalWeight: number;
  OrderItemRate: number;
  ItemPrice: number;
  Amount: number;
  OrderRemarks: string;
  CompCountry: string | null;
  CompContactPerson: string;
  CompMobileA: string;
  CompMobileB: string | null;
  CompMobileC: string | null;
  CompEmailA: string | null;
  CompEmailB: string | null;
  CompLogoImage: string | null;
  CompanyWebsite: string | null;
  CompanyFaxNo: string | null;
  OrgReportingRemarks: string | null;
  OrderItemId: number;
  OrderSupCustId: number;
};
export type TSearchCriteria = {
  FromDate: Date | string;
  ToDate: Date | string;

  ItemId: number;
  Status: string;
  IsApproved: boolean;
  ApprovedStatus: number;
  ApprovedFilter: string;
};